Import 7-zip 4.65
history.txt:

4.65           2009-02-03
-------------------------
- 7-Zip File Manager now can calculate SHA-256 checksum.
- Some bugs were fixed.


4.64           2009-01-03
-------------------------
- The bug in 7-Zip 4.63 was fixed: 7-Zip could not decrypt .ZIP archives
  encrypted with WinZip-AES method.


4.63           2008-12-31
-------------------------
- 7-Zip now can unpack ZIP archives encrypted with PKWARE-AES.
- Some bugs were fixed.


4.62           2008-12-02
-------------------------
- Some bugs were fixed.


4.61 beta      2008-11-23
-------------------------
- 7-Zip now supports LZMA compression for .ZIP archives.
- Some bugs were fixed.
- New localization: Sinhala.


4.60 beta      2008-08-19
-------------------------
- Some bugs were fixed.


4.59 beta      2008-08-13
-------------------------
- 7-Zip now can unpack UDF, XAR and DMG/HFS archives.
- 7-Zip File Manager now keeps encryption when you edit encrypted file inside archive.
- 7-Zip File Manager now allows to change current folder from the address bar drop-down list.
- It's allowed to use -t switch for "list" and "extract" commands.
- Some bugs were fixed.
- New localizations: Icelandic, Kurdish Sorani.


4.58 beta      2008-05-05
-------------------------
- Some speed optimizations.
- 7-Zip now can unpack .lzma archives.
- Unicode (UTF-8) support for filenames in .ZIP archives. Now there are 3 modes:
    1) Default mode: 7-Zip uses UTF-8, if the local code page doesn't contain required symbols.
    2) -mcu switch:  7-Zip uses UTF-8, if there are non-ASCII symbols.
    3) -mcl switch:  7-Zip uses local code page.
- Now it's possible to store file creation time in 7z and ZIP archives (-mtc switch).
- 7-Zip now can unpack multivolume RAR archives created with
  "old style volume names" scheme and names *.001, *.002, ...
- Now it's possible to use -mSW- and -mSW+ switches instead of -mSW=off and -mSW=on
- Some bugs were fixed.
- New localizations: Punjabi (Indian), Pashto.

+++ CPP/7zip/Archive/ArchiveExports.cpp
@@ -13,14 +13,14 @@
 
 static const unsigned int kNumArcsMax = 32;
 static unsigned int g_NumArcs = 0;
-static const CArcInfo *g_Arcs[kNumArcsMax]; 
-void RegisterArc(const CArcInfo *arcInfo) 
-{ 
+static const CArcInfo *g_Arcs[kNumArcsMax];
+void RegisterArc(const CArcInfo *arcInfo)
+{
   if (g_NumArcs < kNumArcsMax)
-    g_Arcs[g_NumArcs++] = arcInfo; 
+    g_Arcs[g_NumArcs++] = arcInfo;
 }
 
-DEFINE_GUID(CLSID_CArchiveHandler, 
+DEFINE_GUID(CLSID_CArchiveHandler,
 0x23170F69, 0x40C1, 0x278A, 0x10, 0x00, 0x00, 0x01, 0x10, 0x00, 0x00, 0x00);
 
 #define CLS_ARC_ID_ITEM(cls) ((cls).Data4[5])
@@ -44,9 +44,9 @@ int FindFormatCalssId(const GUID *clsID)
   if (cls != CLSID_CArchiveHandler)
     return -1;
   Byte id = CLS_ARC_ID_ITEM(*clsID);
-  for (UInt32 i = 0; i < g_NumArcs; i++)
+  for (unsigned i = 0; i < g_NumArcs; i++)
     if (g_Arcs[i]->ClassId == id)
-      return i;
+      return (int)i;
   return -1;
 }
